+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -1802,6 +1802,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       }
 
       @Test
+      @TestMetadata("AssistedFactoryWithContributionProviders.kt")
+      public void testAssistedFactoryWithContributionProviders() {
+        runTest("compiler-tests/src/test/data/box/inject/assisted/AssistedFactoryWithContributionProviders.kt");
+      }
+
+      @Test
       @TestMetadata("AssistedInjectWithConstValQualifiers.kt")
       public void testAssistedInjectWithConstValQualifiers() {
         runTest("compiler-tests/src/test/data/box/inject/assisted/AssistedInjectWithConstValQualifiers.kt");